[tables] Update extract simple table with option to allow gaps
This is because extract_simple_table is much faster than extract_table,
and we often have cases where only a few columns or rows don't have data
in every cell.

Initially we wanted this to only be used on full table, since it can
easily miss cells. We still achieve this when allow_gaps=False
(default), but you can say that you're expecting gaps.