Remove linting requirements from tests
Linting is now handled by pre-commit and pre-commit.ci. Versions for
each of these packages are pinned in .pre-commit-config.yml, and the
versions are periodically updated by pre-commit.ci.

We no longer need these requirements in setup.py. This also resolves an
issue where requirements will get out of sync unless we work to get
dependabot and the pre-commit.ci updates in sync.